fixup! test: split test-runner-watch-mode-kill-signal
