+++ test/common/index.mjs
@@ -46,7 +46,6 @@ const {
   skipIf32Bits,
   skipIfEslintMissing,
   skipIfInspectorDisabled,
-  skipIfNoWatchModeSignals,
   skipIfSQLiteMissing,
   spawnPromisified,
 } = common;
@@ -98,7 +97,6 @@ export {
   skipIf32Bits,
   skipIfEslintMissing,
   skipIfInspectorDisabled,
-  skipIfNoWatchModeSignals,
   skipIfSQLiteMissing,
   spawnPromisified,
 };
