+++ test/parallel/test-watch-mode-kill-signal-override.mjs
@@ -1,13 +1,14 @@
 // Test that the kill signal sent by --watch can be overridden to SIGINT
 // by using --watch-kill-signal.
 
-import { skipIfNoWatchModeSignals } from '../common/index.mjs';
+import '../common/index.mjs';
 import assert from 'node:assert';
 import { writeFileSync } from 'node:fs';
 import { spawn } from 'node:child_process';
 import { once } from 'node:events';
 import tmpdir from '../common/tmpdir.js';
 import fixtures from '../common/fixtures.js';
+import { skipIfNoWatchModeSignals } from '../common/watch.js';
 
 skipIfNoWatchModeSignals();
